% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeListify.R
\name{append_obs}
\alias{append_obs}
\title{This function is a convenience wrapper around largeList::saveList and is 
used to append a new observation to an existing largeList object}
\usage{
append_obs(func_obs, path)
}
\arguments{
\item{func_obs:}{A new functional observation
each observation is a list, that contains two vectors of identical length: args and vals}

\item{path:}{path to where the file shall be saved, including a filename and the 
ending .llo}
}
\value{
no return value
}
\description{
This function is a convenience wrapper around largeList::saveList and is 
used to append a new observation to an existing largeList object
}
