% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparable.R
\name{comparable_obs_finder}
\alias{comparable_obs_finder}
\title{Checks which measuring intervals are suitable to be compared to a given measuring interval.}
\usage{
comparable_obs_finder(main_interval, measuring_intervals, lambda, ids)
}
\arguments{
\item{main_interval:}{vector of two elements: starting and end point of measuring interval}

\item{measuring_intervals:}{use output from measuring_int_mat}

\item{lambda:}{acceptable stretching parameter}

\item{ids:}{identifiers of individual observations}
}
\value{
A list containing two objects: ind, which contains the rownumber of the 
comparable observations in the matrix measuring_intervals and ids, which contains
the identifiers of the comparable observations
}
\description{
Checks which measuring intervals are suitable to be compared to a given measuring interval.
}
