% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{This function can be used to transform the parquet object to a .RDS object
which is necessary for the further steps}
\usage{
data_prep(path, min_length = 0, max_length = 200)
}
\arguments{
\item{path:}{path to the parquet object containing the Endanzug-Data}

\item{min_length:}{minimum length of the endanzug}

\item{max_length:}{maximum permitted length of the endanzug}
}
\value{
placeholder
}
\description{
This function can be used to transform the parquet object to a .RDS object
which is necessary for the further steps
}
