% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_detection.R
\name{detection_wrap}
\alias{detection_wrap}
\title{This function performs the identification procedure implemented in this package
in the simplest scenario. It serves as an important building block for the
more high-level functions.
Important assumptions are the following:
- observations in func_dat are zeroed and have the same measuring interval}
\usage{
detection_wrap(func_dat, ids, alpha, B, gamma = 0.05)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that contains two vectors of identical length: args and vals}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping
(in approximation of C - optional if C is specified)}

\item{ids:}{B: number of smoothed bootstrap samples to use
(in approximation of C - optional if C is specified)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
Returns a list containing two objects:
1. outlier_ids: a vector containing the ids of the observations identified as outliers
2. outlier_ind: a vector containing the position of the the observations identified as outliers in the func_dat list
}
\description{
This function performs the identification procedure implemented in this package
in the simplest scenario. It serves as an important building block for the
more high-level functions.
Important assumptions are the following:
- observations in func_dat are zeroed and have the same measuring interval
}
