% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sets.R
\name{generate_set_3}
\alias{generate_set_3}
\title{Generates a data set containing a chosen number of observations
as described for test set 3}
\usage{
generate_set_3(n_obs = 30000)
}
\arguments{
\item{n_obs:}{number of observations to be generated}
}
\value{
A list of functional observations in the usual format.
}
\description{
Generates a data set containing a chosen number of observations
as described for test set 3
}
