% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{grid_approx_mat}
\alias{grid_approx_mat}
\title{This function is a wrapper around grid_approx_set_obs.
It takes the approximated observations of identical length and transforms 
them into a matrix}
\usage{
grid_approx_mat(func_dat, grid)
}
\arguments{
\item{func_dat:}{list that contains the observations:
each observation is a list, that contains two vectors of identical length: args and vals}

\item{grid:}{grid to use for approximation}
}
\value{
A matrix of dimensions length(func_dat)xlength(grid) that contains 
the approximated values of the functional observations at the grid points
in each row.
}
\description{
This function is a wrapper around grid_approx_set_obs.
It takes the approximated observations of identical length and transforms 
them into a matrix
}
