% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{grid_approx_set_obs}
\alias{grid_approx_set_obs}
\title{This function approximates the functional observations given to it using a piece
wise linear approximation on a specified grid. The format of the data is unchanged}
\usage{
grid_approx_set_obs(func_dat, grid)
}
\arguments{
\item{func_dat:}{list that contains the observations:
each observation is a list, that contains two vectors of identical length: args and vals}

\item{grid:}{grid to use for approximation}
}
\value{
A list of identical length to func_dat. Each entry corresponds to the 
same entry in func_dat but is an approximation using linear interpolation on
the points given in grid.
}
\description{
This function approximates the functional observations given to it using a piece
wise linear approximation on a specified grid. The format of the data is unchanged
}
