% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{grid_finder}
\alias{grid_finder}
\title{This function creates a grid that is used for the piece wise linear
approximation of functions. Currently this assumes that the observations
share a measuring interval and constructs an equdistant grid of specified
length.}
\usage{
grid_finder(func_dat, length.out = 100)
}
\arguments{
\item{func_dat:}{Data used by functional procedures}

\item{length.out:}{number of points in the grid}
}
\value{
A vector that contains the grid points.
}
\description{
This function creates a grid that is used for the piece wise linear
approximation of functions. Currently this assumes that the observations
share a measuring interval and constructs an equdistant grid of specified
length.
}
