% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeListify.R
\name{largeListify}
\alias{largeListify}
\title{This function is a convenience wrapper around largeList::saveList}
\usage{
largeListify(func_dat, path)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that contains two vectors of identical length: args and vals}

\item{path:}{path to where the file shall be saved, including a filename and the 
ending .llo}
}
\value{
Returns NULL.
}
\description{
This function is a convenience wrapper around largeList::saveList
}
