% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervals.R
\name{measuring_int}
\alias{measuring_int}
\title{This function extracts the measuring interval from a functional observation.}
\usage{
measuring_int(func_obs)
}
\arguments{
\item{func_obs:}{a list, that contains two vectors of identical length: args and vals}
}
\value{
A vector of two elements, the start and endpoints of the measuring interval
}
\description{
This function extracts the measuring interval from a functional observation.
}
