% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervals.R
\name{measuring_int_list}
\alias{measuring_int_list}
\title{This function extracts the measuring interval from a set of functional observations.}
\usage{
measuring_int_list(func_dat)
}
\arguments{
\item{func_dat:}{a list, that contains two vectors of identical length: args and vals}
}
\value{
A list of vectors containing two elements, the start and endpoints of the measuring intervals
}
\description{
This function extracts the measuring interval from a set of functional observations.
}
