% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervals.R
\name{measuring_int_mat}
\alias{measuring_int_mat}
\title{This function is a wrapper around measuring_int_list and combines the intervals
into a length(func_dat)x2 matrix}
\usage{
measuring_int_mat(func_dat)
}
\arguments{
\item{func_dat:}{a list, that contains two vectors of identical length: args and vals}
}
\value{
A length(func_dat)x2 matrix that contains in each row the measuring interval
of the corresponding functional observation.
}
\description{
This function is a wrapper around measuring_int_list and combines the intervals
into a length(func_dat)x2 matrix
}
