% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedure.R
\name{outlier_detection}
\alias{outlier_detection}
\title{placeholder}
\usage{
outlier_detection(
  matr_dat,
  alpha = 0.05,
  B = 100,
  gamma = 0.05,
  ids,
  grid,
  C = NULL
)
}
\arguments{
\item{matr_dat:}{data in matrix form - each row contains the grid approximations of one observation}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping
(in approximation of C - optional if C is specified)}

\item{B:}{number of smoothed bootstrap samples to use
(in approximation of C - optional if C is specified)}

\item{gamma:}{tuning parameter for smoothed bootstrap}

\item{ids:}{identifiers of individual observations}

\item{grid:}{grid used in approximation of matr_dat}

\item{C:}{should be provided. Otherwise C will be approximated in each step of the iteration}
}
\value{
placeholder
}
\description{
placeholder
}
