% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedure.R
\name{outlier_iteration}
\alias{outlier_iteration}
\title{This function performs one iteration of the algorithm, including the 
calculation of functional depths, the approximation of C and the 
flagging of observations with depths lower than C}
\usage{
outlier_iteration(matr_dat, alpha = 0.05, B = 50, gamma, ids, grid, C = NULL)
}
\arguments{
\item{matr_dat:}{data in matrix form - each row contains the grid approximations of one observation}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping
(in approximation of C - optional if C is specified)}

\item{:}{C: should be provided. Otherwise C will be approximated in each step of the iteration}
}
\value{
A list containing three objects:
matr_data: a reduced matrix containing the observations not classified as outliers
ids: the ids corresponding to the observations not classified as outliers
outlier_ids: the ids corresponding to the observations classified as outliers
}
\description{
This function performs one iteration of the algorithm, including the 
calculation of functional depths, the approximation of C and the 
flagging of observations with depths lower than C
}
