% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{possible_occurences}
\alias{possible_occurences}
\title{This function decides which measuring intervals are suitable to be compared 
to a given new observation}
\usage{
possible_occurences(func_obs, unique_intervals, lambda)
}
\arguments{
\item{func_obs:}{list with two elements: vectors of equal length called args and vals}

\item{unique_intervals:}{matrix containing measuring intervals that occur in the data set (one in each row)}

\item{lambda:}{acceptable stretching parameter}
}
\value{
A list containing two elements
occurs: the indices (rownumbers) of the intervals that are comparable 
to the given functional observation
occurs_intervals: the matrix containing only the comparable intervals
}
\description{
This function decides which measuring intervals are suitable to be compared 
to a given new observation
}
