% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sets.R
\name{random_dat_1}
\alias{random_dat_1}
\title{Placeholder}
\usage{
random_dat_1(grid, slope, out)
}
\arguments{
\item{grid:}{Grid where the observations take palce}

\item{slope:}{Slope parameter for the generated curves}

\item{out:}{Vector conatining zeroes where no outlier should be generated 
and ones where an outlier is generated}
}
\value{
Placeholder
}
\description{
Placeholder
}
