% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sets.R
\name{random_dat_2}
\alias{random_dat_2}
\title{Generates one observation for test set 2}
\usage{
random_dat_2(grid, slope, out)
}
\arguments{
\item{grid:}{Grid where the observations take palce}

\item{slope:}{Slope parameter for the generated curves}

\item{out:}{Vector containing zeroes where no outlier should be generated}
}
\value{
A functional observation in the usual format
}
\description{
Generates one observation for test set 2
}
