% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_wrap.R
\name{sample_helper}
\alias{sample_helper}
\title{placeholder}
\usage{
sample_helper(list_path, ids, index, alpha, B, gamma)
}
\arguments{
\item{list_path:}{path to the random access list of the data set (generated by package largeList)}

\item{index:}{index of observations to use in the procedure}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping (in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
placeholder
}
\description{
placeholder
}
