% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_wrap.R
\name{sample_wrap}
\alias{sample_wrap}
\title{Performs the sampling procedure on a data set.}
\usage{
sample_wrap(
  cl,
  list_path,
  indeces = NULL,
  n_samples,
  sample_size,
  alpha,
  B,
  gamma
)
}
\arguments{
\item{cl:}{cluster object generated by parallel package}

\item{list_path:}{path to the random access list of the data set (generated by package largeList)
assumes that the observations are already zeroed and share the same measuring
interval}

\item{indices:}{Vector containing the indeces of observations that are
included in the sampling procedure}

\item{n_samples:}{number of samples to use}

\item{sample_size:}{number of observations to use in each sample}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping
(in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
A list containing three elements:
num_samples: The number of samples each observation was part of (vector)
num_outliers: The number of samples each observation was classified as an outlier in (vector)
certainties: The quotient of num_outliers and num_samples (so the relative frequency of 
observations being classified as outliers)
}
\description{
Performs the sampling procedure on a data set.
}
