% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_auxiliary.R
\name{sampling_number}
\alias{sampling_number}
\title{This function determines}
\usage{
sampling_number(n_comparables, sample_size, expn)
}
\arguments{
\item{n_comparables:}{Number of comparable observations}

\item{sample_size:}{Chosen sample size for sampling procedure}

\item{expn:}{Chosen expected number of draws any observation appears in per 
unique interval it is comparable to.}
}
\value{
A number that's used as the number of samples drawn in each
sampling step.
}
\description{
This function determines
}
