% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stretch.R
\name{stretch_data}
\alias{stretch_data}
\title{This function is a wrapper around stretch_obs. It takes a set of functional 
observations func_dat and stretches each observation to fit a given measuring interval}
\usage{
stretch_data(func_dat, measuring_interval)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that contains two vectors of identical length: args and vals}

\item{measuring_interval:}{a vector with 2 elements,
the start and end points of the desired measuring interval}
}
\value{
placeholder
}
\description{
This function is a wrapper around stretch_obs. It takes a set of functional 
observations func_dat and stretches each observation to fit a given measuring interval
}
