% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stretch.R
\name{stretch_obs}
\alias{stretch_obs}
\title{This function takes a functional observation func_obs and stretches it to 
fit a given measuring interval}
\usage{
stretch_obs(func_obs, measuring_interval)
}
\arguments{
\item{func_obs:}{a list that contains two vectors of identical length: args and vals}

\item{measuring_interval:}{a vector with 2 elements,
the start and end points of the desired measuring interval}
}
\value{
placeholder
}
\description{
This function takes a functional observation func_obs and stretches it to 
fit a given measuring interval
}
