% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch_sample_wrap.R
\name{stretch_sample_helper}
\alias{stretch_sample_helper}
\title{This function is a helper to stretch_sample_wrap() and performs the detection procedure for
a single sample.}
\usage{
stretch_sample_helper(list_path, main_interval, ids, index, alpha, B, gamma)
}
\arguments{
\item{list_path:}{path to the random access list of the data set (generated by package largeList)}

\item{main_interval:}{Measuring interval currently under consideration}

\item{index:}{index of observations to use in the procedure 
(decided to be comparable in the calling function)}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping (in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
The output from detection_wrap for further usage.
}
\description{
This function is a helper to stretch_sample_wrap() and performs the detection procedure for
a single sample.
}
