% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{stretch_sample_updating}
\alias{stretch_sample_updating}
\title{placeholder}
\usage{
stretch_sample_updating(
  cl,
  new_observation,
  list_path,
  lambda,
  measuring_intervals,
  sample_size,
  expn,
  alpha,
  B,
  gamma,
  num_samples_prev,
  num_outliers_prev,
  debug = FALSE
)
}
\arguments{
\item{cl:}{cluster object generated by parallel package}

\item{new_observation:}{new observations added to the data set}

\item{list_path:}{path to the random access list of the data set (generated by package largeList)
this assumes that all observations are already zeroed and for every observation
there are more comparable counterparts than the parameter sample_size}

\item{lambda:}{allowed stretching parameter}

\item{n_samples:}{number of samples to use}

\item{sample_size:}{number of observations to use in each sample}

\item{expn:}{Chosen expected number of draws any observation appears in per
unique interval it is comparable to.}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping
(in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}

\item{num_samples_prev:}{value of num_samples before new observation was added}

\item{num_outliers_prev:}{value of num_outliers before new observation was added}

\item{debug:}{choose whether text outputs are given}
}
\value{
placeholder
}
\description{
placeholder
}
