% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch_sample_wrap.R
\name{stretch_sample_wrap}
\alias{stretch_sample_wrap}
\title{placeholder}
\usage{
stretch_sample_wrap(
  cl,
  list_path,
  measuring_interval,
  indeces = NULL,
  n_samples,
  sample_size,
  alpha,
  B,
  gamma
)
}
\arguments{
\item{cl:}{cluster object generated by parallel package}

\item{list_path:}{path to the random access list of the data set (generated by package largeList)}

\item{measuring_itnerval:}{measuring interval the observations are stretched to}

\item{indices:}{Vector containing the indeces of observations that are
included in the sampling procedure}

\item{n_samples:}{number of samples to use}

\item{sample_size:}{number of observations to use in each sample}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping
(in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
placeholder
}
\description{
placeholder
}
