% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{stsa_update_helper}
\alias{stsa_update_helper}
\title{Performs the updating procedure for a single sample.}
\usage{
stsa_update_helper(
  list_path,
  new_observation,
  new_id,
  main_interval,
  ids,
  index,
  alpha,
  B,
  gamma
)
}
\arguments{
\item{list_path:}{path to the random access list of the data set (generated by package largeList)}

\item{new_observation:}{new observations added to the data set}

\item{new_id:}{identifier for the new observations added to the data set}

\item{main_interval:}{Measuring interval currently under consideration}

\item{index:}{index of observations to use in the procedure
(decided to be comparable in the calling function)}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping (in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
The outlier_ids part from the detection_wrap function.
}
\description{
Performs the updating procedure for a single sample.
}
