% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervals.R
\name{unique_intervals}
\alias{unique_intervals}
\title{Given a matrix of measuring intervals as generated by measuring_int_mat
this function extracts the unique intervals in the data set and returns
a reduced matrix containing only the unique intervals}
\usage{
unique_intervals(interval_matrix)
}
\arguments{
\item{interval_matrix:}{A matrix with 2 columns containing in its rows
the measuring intervals of the observations}
}
\value{
A matrix containing the unique measuring intervals found 
in the data set
}
\description{
Given a matrix of measuring intervals as generated by measuring_int_mat
this function extracts the unique intervals in the data set and returns
a reduced matrix containing only the unique intervals
}
