% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zeroing.R
\name{zero_data}
\alias{zero_data}
\title{This function is a wrapper around zero_obs and zeroes all 
observations in the data set}
\usage{
zero_data(func_dat)
}
\arguments{
\item{func_dat:}{list that contains the observations
each observation is a list, that contains two vectors of identical length: args and vals}
}
\value{
A list of zeroed observations
}
\description{
This function is a wrapper around zero_obs and zeroes all 
observations in the data set
}
