% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zeroing.R
\name{zero_obs}
\alias{zero_obs}
\title{This function zeroes an observation by shifting the arg values to begin at zero.}
\usage{
zero_obs(func_obs)
}
\arguments{
\item{func_obs:}{a list, that contains two vectors of identical length: args and vals}
}
\value{
a list, that contains two vectors of identical length: args and vals,
where args begins with zero.
}
\description{
This function zeroes an observation by shifting the arg values to begin at zero.
}
