% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sets.R
\name{generate_set_3}
\alias{generate_set_3}
\title{Placeholder}
\usage{
generate_set_3(n_obs = 30000)
}
\arguments{
\item{n_obs:}{number of observations to be generated}
}
\value{
Placeholder
}
\description{
Placeholder
}
