% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{possible_occurences}
\alias{possible_occurences}
\title{Placeholder}
\usage{
possible_occurences(func_obs, unique_intervals, lambda)
}
\arguments{
\item{func_obs:}{list with two elements: vectors of equal length called args and vals}

\item{unique_intervals:}{matrix containing measuring intervals that occur in the data set (one in each row)}

\item{lambda:}{acceptable stretching parameter}
}
\description{
Placeholder
}
