% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch_sample_wrap.R
\name{stretch_sample_helper}
\alias{stretch_sample_helper}
\title{placeholder}
\usage{
stretch_sample_helper(
  list_path,
  measuring_interval,
  ids,
  index,
  alpha,
  B,
  gamma
)
}
\arguments{
\item{list_path:}{path to the random access list of the data set (generated by package largeList)}

\item{measuring_interval:}{Measuring interval currently under consideration}

\item{index:}{index of observations to use in the procedure 
(decided to be comparable in the calling function)}

\item{alpha:}{quantile of least deep observations to drop before bootstrapping (in approximation of C)}

\item{B:}{number of smoothed bootstrap samples to use (in approximation of C)}

\item{gamma:}{tuning parameter for smoothed bootstrap}
}
\value{
placeholder
}
\description{
placeholder
}
